Update ProjectContextBuilder and FrameworkReferenceResolver
