+++ src/Microsoft.DotNet.ProjectModel/ProjectContextBuilder.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
 using Microsoft.DotNet.ProjectModel.Graph;
 using Microsoft.DotNet.ProjectModel.Resolution;
 using Microsoft.Extensions.Internal;
@@ -154,7 +153,7 @@ namespace Microsoft.DotNet.ProjectModel
 
             RootDirectory = GlobalSettings?.DirectoryPath ?? RootDirectory;
             PackagesDirectory = PackagesDirectory ?? PackageDependencyProvider.ResolvePackagesPath(RootDirectory, GlobalSettings);
-            ReferenceAssembliesPath = ReferenceAssembliesPath ?? GetDefaultReferenceAssembliesPath();
+            ReferenceAssembliesPath = ReferenceAssembliesPath ?? FrameworkReferenceResolver.GetDefaultReferenceAssembliesPath();
 
             LockFileLookup lockFileLookup = null;
 
@@ -345,44 +344,6 @@ namespace Microsoft.DotNet.ProjectModel
 
                 libraries.Add(new LibraryKey(library.Name), description);
             }
-        }
-
-        public static string GetDefaultReferenceAssembliesPath()
-        {
-            // Allow setting the reference assemblies path via an environment variable
-            var referenceAssembliesPath = Environment.GetEnvironmentVariable("DOTNET_REFERENCE_ASSEMBLIES_PATH");
-
-            if (!string.IsNullOrEmpty(referenceAssembliesPath))
-            {
-                return referenceAssembliesPath;
-            }
-
-            if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                // There is no reference assemblies path outside of windows
-                // The enviorment variable can be used to specify one
-                return null;
-            }
-
-            // References assemblies are in %ProgramFiles(x86)% on
-            // 64 bit machines
-            var programFiles = Environment.GetEnvironmentVariable("ProgramFiles(x86)");
-
-            if (string.IsNullOrEmpty(programFiles))
-            {
-                // On 32 bit machines they are in %ProgramFiles%
-                programFiles = Environment.GetEnvironmentVariable("ProgramFiles");
-            }
-
-            if (string.IsNullOrEmpty(programFiles))
-            {
-                // Reference assemblies aren't installed
-                return null;
-            }
-
-            return Path.Combine(
-                    programFiles,
-                    "Reference Assemblies", "Microsoft", "Framework");
         }
 
         private void EnsureProjectLoaded()
